DFG-in-egraph: add a new ValueDef node kind: a union of two other nodes.
This commit adds a new kind of value that is a union of two other
values. It is not yet constructed, supported in CLIF text format, or
used in any way, but it is the basis for representing an
aegraph (acyclic egraph) in the CLIF `DataFlowGraph` data structure
rather than a separate thing.