Egraph-in-DFG: implement stripping-to-the-bones of the function body.
In order to do processing on the egraph, we need to remove all "pure"
nodes (for some careful definition of pure) from the function layout,
so that we can reason about these operators as values and place them
later. We keep the "side-effect skeleton" around, and remove all other
insts.