Cranelift: ISLE: fix overly-recursive rule on selects-of-truthy-values.
Our rules contain some simplifications that optimize `select`
operators when the selector input is one of several "truthy"
operators; that is, operators, that preserve zero/nonzero (without
necessarily preserving the bits), such as rotates.

In one input found by a fuzzbug, this rule led to quadtratic behavior
during compilation with the new aegraph internals that canonicalize
less. The core issue was that the rule added `(simplify x t f)` to an
eclass with `(simplify (rotr x ...) t f)`; but on creating the new
node, triggered a recursive rewrite, which in turn re-entered the
rule. (This is `v224` in the included test-case.)

I would like to study a bit more the precise way in which full
canonicalization previously prevented this, but making the rule
subsume rather than union into the eclass prevents the quadratic
blowup.